+++ functions.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * Theme bootstrap.
+ * Theme bootstrap for LOOPIS sub sites (aka. local apps)
  *
  * Loads all frontend core files.
  */
@@ -16,14 +16,11 @@ if (!defined('ABSPATH')) { exit; }
  */
 
 // Define theme version
-define('LOOPIS_THEME_VERSION', '0.84'); // Update version number here + in style.css
+define('LOOPIS_THEME_VERSION', '0.86'); // Update version number here + in style.css
 
 // Define theme folder path constants
 define('LOOPIS_THEME_DIR', get_template_directory());       // Server-side path to /wp-content/themes/loopis-theme/
 define('LOOPIS_THEME_URI', get_template_directory_uri());   // Client-side path to https://loopis.app/wp-content/themes/loopis-theme/
-
-// Define secret keys from .env file (if not live)
-require_once __DIR__ . '/environment_loader.php';
 
 // Define locker ID for this installation (temporary solution)
 define('LOCKER_ID', '12845-1');
@@ -72,6 +69,7 @@ function loopis_theme_load_files() {
     loopis_theme_include_folder('shortcodes');
     loopis_theme_include_folder('filters');
     loopis_theme_include_folder('functions/everyone');
+    loopis_theme_include_folder('functions/payment');
 
     // For user
     if (is_user_logged_in()) {