+++ templates/faq/questions-visitor.php
@@ -12,7 +12,7 @@ if (!defined('ABSPATH')) {
 $faq_url = home_url( '/faq/' );
 ?>
 
-<div class="columns"><div class="column1"><h5>💡 Vanliga frågor</h5></div>
+<div class="columns"><div class="column1"><h3>💡 Vanliga frågor</h3></div>
 <div class="column2 bottom"><a href="<?php echo esc_url($faq_url);?>">Alla frågor & svar →</a></div></div>
 <hr>
 <p><span class="big-link"><a href="<?php echo esc_url($faq_url . 'hur-funkar-loopis/');?>">📌 Hur funkar LOOPIS?</a></span></p>