+++ templates/user/front-page/add-to-homescreen.php
@@ -9,11 +9,11 @@ if (!defined('ABSPATH')) {
 }
 ?>
 
-<div id="homescreen" style="display: none; position: fixed; height: 20px; width: 100%; bottom: 80px; text-align: center; background-color: #fff; border: 1px solid #e5e5e5; padding: 5px; z-index: 9998;">
+<div id="homescreen" class="loopis-message notification" style="display: none; position: fixed; height: 20px; width: 100%; bottom: 80px; text-align: center; background-color: #fff; border: 1px solid #e5e5e5; padding: 5px; z-index: 9998;">
     <p class="small">💡<a href="#" onclick="showArrow(); return false;"> Lägg LOOPIS på hemskärmen</a> så liknar det en app!</p>
 </div>
 
-<div id="arrow" class="wpum-message notification" style="display: none; position: fixed; bottom: 0; left: 50%; transform: translateX(-50%); z-index: 9999; text-align: center;">
+<div id="arrow" class="loopis-message notification" style="display: none; position: fixed; bottom: 0; left: 50%; transform: translateX(-50%); z-index: 9999; text-align: center;">
     <p>Tryck på dela-knappen i <b>Safari</b> och scrolla ner till <b>Lägg till på hemskärmen</b></p>
     <img style="width: 72px;" src="<?php echo get_template_directory_uri(); ?>/assets/img/homescreen-iphone.png" alt="arrow" />
     <a style="position: absolute; right: 20px; bottom: 10px;" href="#" onclick="hideArrow(); return false;"><i class="fas fa-times"></i> Stäng</a>